delete album with connections from logged user

+++ backend/src/main/java/pl/polsl/snapsort/controller/AlbumController.java
@@ -14,6 +14,7 @@ import pl.polsl.snapsort.service.JwtTokenUtil;
 import pl.polsl.snapsort.service.PhotoService;
 
 import javax.persistence.EntityNotFoundException;
+import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
 
@@ -124,10 +125,19 @@ public class AlbumController {
         }
     }
 
-
+    @Transactional
     @DeleteMapping("/{id}")
-    public void deleteAlbum(@PathVariable Long id) {
+    public void deleteAlbum(@PathVariable Long id, @RequestHeader("Authorization") String token) {
+        Long userId = jwtTokenUtil.extractUserId(token.replace("Bearer ", ""));
+
+        // Check if the album belongs to the logged-in user
+        boolean albumBelongsToUser = albumService.existsAlbumByIdAndUserId(id, userId);
+        if (!albumBelongsToUser) {
+            throw new AlbumNotFoundException("Album not found or does not belong to the user.");
+        }
+
         albumService.deleteAlbum(id);
     }
+
 }
